requirements.txt
nltk
spacy
scikit-learn
sentence-transformers
numpy
requests
google-genai
